@@ -4,6 +4,7 @@
   "description": "GitHub Action to generate changelog from conventional commits",
   "main": "dist/index.js",
   "scripts": {
+    "lint": "eslint .",
     "build": "ncc build index.js -o dist"
   },
   "repository": {
@@ -22,17 +23,18 @@
   },
   "homepage": "https://github.com/Requarks/changelog-action#readme",
   "dependencies": {
-    "@actions/core": "1.6.0",
-    "@actions/github": "5.0.0",
+    "@actions/core": "1.10.0",
+    "@actions/github": "5.1.1",
     "@conventional-commits/parser": "0.4.1",
     "lodash": "4.17.21"
   },
   "devDependencies": {
-    "@vercel/ncc": "0.33.1",
-    "eslint": "8.7.0",
-    "eslint-config-standard": "16.0.3",
-    "eslint-plugin-import": "2.25.4",
+    "@vercel/ncc": "0.34.0",
+    "eslint": "8.26.0",
+    "eslint-config-standard": "17.0.0",
+    "eslint-plugin-n": "15.3.0",
+    "eslint-plugin-import": "2.26.0",
     "eslint-plugin-node": "11.1.0",
-    "eslint-plugin-promise": "6.0.0"
+    "eslint-plugin-promise": "6.1.1"
   }
 }
